submit of lemin to hive
