--- Issues
@@ -1,29 +1,0 @@
-visualizer segfault, because search comes before map data.
-
-Clean .h file. Comments.
-
-memory leaks
-
-free memory in error cases
-
-input not freed when no paths found.
-
-maps w leaks
-invalid/links_no_links
-invalid/links/no_path
-invalid/map/invalidcommand1
-maps/invalid/map/map_loop1
-invalid/map/no_tubes
-
-
-invalid/map_map_size_1000
-invalid/map/start_is_end
-invali/map/start_not_followed_by_room
-invalid/rooms/room_no_y_coord
-invalid/rooms/room_too_much_arg
-invalid/rooms/withot_coord
-
-after make re libft there is collect2: error
-
-
-testing
