+++ tests/map_f1
@@ -1,0 +1,443 @@
+1
+#Here is the number of lines required: 36
+##start
+Qtj1 2 2
+##end
+Pbf1 5 5
+Ylq4 7 7
+Hd_0 9 9
+Nvg8 11 11
+Rrv7 14 14
+Cql1 16 16
+Xx_8 20 20
+Nyt2 23 23
+N_c0 27 27
+Evu6 29 29
+Gxg8 34 34
+U_f7 37 37
+Cza5 41 41
+Yhf1 43 43
+Mel5 46 46
+Od_9 51 51
+Gyr3 53 53
+Jjk6 56 56
+Qpk4 61 61
+Qdr2 65 65
+Vzr6 67 67
+Tvx9 72 72
+Swq0 75 75
+Bj_9 80 80
+Mtj6 84 84
+Y_p1 88 88
+Mzb1 91 91
+Oyq8 95 95
+Wom3 96 96
+F_k2 97 97
+Hjv6 99 99
+Egj7 103 103
+Nur3 104 104
+Zxj4 106 106
+Qli9 107 107
+Fgm3 109 109
+Wxl8 113 113
+Mqf3 115 115
+Wix1 118 118
+Esf9 122 122
+Rsv8 126 126
+Pyl5 128 128
+Daa7 130 130
+I_j8 132 132
+Fqb2 136 136
+Bum9 139 139
+F_s1 142 142
+Fku4 145 145
+Kfi0 146 146
+Qen0 150 150
+Pli8 154 154
+Nbx2 156 156
+X_v3 157 157
+Pbm4 159 159
+S__7 162 162
+Dxl8 167 167
+Gry5 170 170
+Pdi0 172 172
+H_o3 175 175
+Eph1 177 177
+Fkv0 182 182
+Vbd3 187 187
+E_c6 189 189
+Ed_1 193 193
+Xz_9 196 196
+J_a4 198 198
+Tj_2 202 202
+Wdx2 205 205
+Vpv1 208 208
+Nna1 213 213
+Ydm3 214 214
+Dci6 217 217
+Ttf1 220 220
+Mzj2 223 223
+Clg7 226 226
+Mpl0 229 229
+X_s6 232 232
+Jqc3 237 237
+Bhv4 241 241
+Ne_6 245 245
+Qbd6 249 249
+Kn_6 252 252
+Tnq5 255 255
+Dpg3 259 259
+Rco6 263 263
+Ny_8 265 265
+Rvz5 268 268
+Rqa3 272 272
+Haa6 274 274
+Nue1 278 278
+Gxt3 282 282
+Kp_6 285 285
+Fhc4 288 288
+Ovc1 291 291
+Hj_9 295 295
+Wa_6 298 298
+Ail0 300 300
+Npl1 302 302
+Xvr9 303 303
+Myw7 304 304
+Htt1 306 306
+I_u1 307 307
+Qgu6 312 312
+Xma7 314 314
+Fti2 317 317
+Trs7 321 321
+Mbz7 325 325
+Rxx7 327 327
+Jfi1 335 335
+Py_5 338 338
+Rof7 341 341
+Hcc4 343 343
+El_4 346 346
+In_7 347 347
+Rct7 350 350
+U_m2 353 353
+Fp_3 355 355
+Ygl1 359 359
+Jqb6 363 363
+Jgb7 366 366
+T_m5 368 368
+Jwu4 369 369
+Nre7 370 370
+Clp4 373 373
+Mfm1 375 375
+Gpy8 377 377
+Cec7 378 378
+Ctb4 381 381
+Yky9 384 384
+Yzh5 389 389
+Dax2 393 393
+Xxj7 397 397
+Ofz6 401 401
+Vad5 406 406
+Bqv1 409 409
+Ybj3 412 412
+Afc1 413 413
+Wgr3 417 417
+Zvd7 420 420
+Sfc6 422 422
+Xqk1 425 425
+Wmu3 430 430
+Bo_0 432 432
+Wna9 435 435
+Pof7 438 438
+Vec0 440 440
+Nzr5 443 443
+Khc0 447 447
+Dw_2 449 449
+Tgz8 451 451
+Ewa3 453 453
+Hdn3 455 455
+Avz9 459 459
+Ujz0 462 462
+Mdv8 467 467
+K_o6 471 471
+Sry1 475 475
+G__6 479 479
+Kgg6 482 482
+Igg4 484 484
+Eyl7 489 489
+Uxt1 491 491
+Jkm1 494 494
+Zum5 497 497
+V_a9 501 501
+Ubo8 502 502
+Hp_8 504 504
+Jlu3 507 507
+R_n3 510 510
+M__3 512 512
+Xha7 517 517
+Jeb8 519 519
+Piw3 521 521
+Npe7 524 524
+Cfa3 529 529
+Fq_4 534 534
+M_b8 536 536
+K_y1 539 539
+Rbx5 542 542
+Yip1 546 546
+I_p3 548 548
+Ght6 552 552
+Hwx5 556 556
+B_t2 558 558
+Ss_5 563 563
+Xxq6 564 564
+Bs_2 567 567
+Jwk4 570 570
+Ejm2 574 574
+Cra5 577 577
+Sqj0 580 580
+Atf2 583 583
+Mkb3 584 584
+Rlw6 586 586
+Okz1 588 588
+Olp5 593 593
+Kym8 595 595
+Hny6 598 598
+Kyh4 599 599
+Yst3 600 600
+Nsw3 603 603
+Shs9 608 608
+Upn5 609 609
+Ers0 610 610
+Chf6 614 614
+Gt_5 617 617
+Kv_5 621 621
+Qov3 626 626
+Vci0 627 627
+Tcd4 628 628
+Xid5 630 630
+Ca_1 634 634
+Mwp7 635 635
+Xgg6 638 638
+Yso6 641 641
+Cds2 644 644
+Gv_1 647 647
+Ecc0 652 652
+In_7-Hd_0
+Esf9-Ecc0
+Cec7-Cds2
+Bo_0-Rrv7
+Kym8-Xgg6
+B_t2-Xid5
+Rco6-Xx_8
+Jjk6-Vci0
+Xxj7-N_c0
+Mdv8-Qov3
+Nna1-Chf6
+Fq_4-Ers0
+Cra5-Upn5
+Rvz5-Nsw3
+Dax2-Kyh4
+Rlw6-Hny6
+Swq0-Kym8
+Myw7-Olp5
+Qbd6-Mkb3
+Clg7-Gxg8
+In_7-U_f7
+Hdn3-Sqj0
+F_k2-Yhf1
+Tvx9-Mel5
+Jeb8-Jwk4
+X_s6-Bs_2
+Tj_2-Ss_5
+Trs7-Qpk4
+Ujz0-Hwx5
+S__7-Ght6
+Wdx2-Vzr6
+Jkm1-Yip1
+Oyq8-Tvx9
+Mtj6-Swq0
+Wa_6-Rbx5
+Wom3-Bj_9
+J_a4-Y_p1
+Zum5-Mzb1
+Dci6-K_y1
+Zum5-Fq_4
+Nur3-Egj7
+Jeb8-Zxj4
+Pli8-Piw3
+Yky9-Qli9
+Nbx2-Jeb8
+R_n3-Xha7
+Pli8-Fgm3
+M__3-Wxl8
+Esf9-M__3
+Rsv8-R_n3
+Jlu3-Rsv8
+Hp_8-Jlu3
+Ubo8-Hp_8
+Jkm1-Daa7
+Eyl7-Uxt1
+I_j8-Eyl7
+Fqb2-Igg4
+Bum9-Fqb2
+F_s1-Kgg6
+Sry1-G__6
+K_o6-Sry1
+Mdv8-Qen0
+Ujz0-Mdv8
+Avz9-Pli8
+Nbx2-Avz9
+Dxl8-S__7
+Gry5-Dxl8
+Fkv0-Eph1
+Ewa3-Vbd3
+Dw_2-Tgz8
+Tj_2-J_a4
+Nna1-Vpv1
+Dci6-Wna9
+Ttf1-Bo_0
+Clg7-Mzj2
+Wmu3-Clg7
+Xqk1-Wmu3
+Mpl0-Xqk1
+X_s6-Mpl0
+Zvd7-Sfc6
+Jqc3-Zvd7
+Bhv4-Jqc3
+Afc1-Wgr3
+Qbd6-Ne_6
+Ybj3-Qbd6
+Bqv1-Ybj3
+Vad5-Tnq5
+Dpg3-Vad5
+Rco6-Dpg3
+Haa6-Rqa3
+Nue1-Haa6
+Ofz6-Gxt3
+Xxj7-Ofz6
+Dax2-Xxj7
+Kp_6-Dax2
+Ovc1-Fhc4
+Qtj1-Ovc1
+Jfi1-Pbf1
+Ctb4-Yky9
+Hj_9-Ctb4
+Wa_6-Hj_9
+Ail0-Gpy8
+Mfm1-Ail0
+Clp4-Npl1
+Xvr9-Nre7
+Jwu4-Xvr9
+T_m5-Myw7
+I_u1-Jqb6
+Ygl1-I_u1
+Rct7-U_m2
+Xma7-In_7
+El_4-Fti2
+Hcc4-Mbz7
+Rxx7-Hcc4
+Py_5-Rof7
+Qtj1-Py_5
+Rof7-Rxx7
+Mbz7-Trs7
+Trs7-El_4
+Fti2-Xma7
+In_7-Qgu6
+Qgu6-Rct7
+U_m2-Fp_3
+Fp_3-Ygl1
+Jqb6-Htt1
+Htt1-Jgb7
+Jgb7-T_m5
+Myw7-Jwu4
+Nre7-Clp4
+Npl1-Mfm1
+Gpy8-Cec7
+Cec7-Wa_6
+Yky9-Jfi1
+Fhc4-Yzh5
+Yzh5-Kp_6
+Gxt3-Nue1
+Rqa3-Rvz5
+Rvz5-Ny_8
+Ny_8-Rco6
+Tnq5-Kn_6
+Kn_6-Bqv1
+Ne_6-Afc1
+Wgr3-Bhv4
+Sfc6-X_s6
+Mzj2-Ttf1
+Bo_0-Dci6
+Wna9-Ydm3
+Ydm3-Pof7
+Pof7-Nna1
+Vpv1-Vec0
+Vec0-Wdx2
+Wdx2-Nzr5
+Nzr5-Tj_2
+J_a4-Xz_9
+Xz_9-Khc0
+Khc0-Dw_2
+Tgz8-Ed_1
+Ed_1-E_c6
+E_c6-Ewa3
+Vbd3-Fkv0
+Eph1-H_o3
+H_o3-Hdn3
+Hdn3-Pdi0
+Pdi0-Gry5
+S__7-Pbm4
+Pbm4-X_v3
+X_v3-Nbx2
+Pli8-Ujz0
+Qen0-Kfi0
+Kfi0-Fku4
+Fku4-K_o6
+G__6-F_s1
+Kgg6-Bum9
+Igg4-I_j8
+Uxt1-Jkm1
+Daa7-Zum5
+Zum5-V_a9
+V_a9-Pyl5
+Pyl5-Ubo8
+R_n3-Esf9
+M__3-Wix1
+Wix1-Pbf1
+I_u1-Mqf3
+Fp_3-Nur3
+Rxx7-Npe7
+G__6-Cfa3
+Bqv1-M_b8
+Xxj7-Hjv6
+Kgg6-F_k2
+Zvd7-Wom3
+V_a9-Oyq8
+U_m2-Mtj6
+Qbd6-I_p3
+S__7-B_t2
+V_a9-Qdr2
+Yzh5-Jjk6
+J_a4-Gyr3
+Vzr6-Xxq6
+Xha7-Od_9
+I_u1-Ejm2
+Dci6-Cra5
+Ubo8-Cza5
+Mpl0-Atf2
+Nur3-Rlw6
+Ybj3-Okz1
+Afc1-Evu6
+Sqj0-Yst3
+Ttf1-Shs9
+Nur3-Gt_5
+U_m2-Kv_5
+Kgg6-Nyt2
+Zxj4-Tcd4
+Rlw6-Ca_1
+Fgm3-Mwp7
+Khc0-Cql1
+Daa7-Nvg8
+Fku4-Yso6
+Rct7-Gv_1
+Cfa3-Ylq4
+#Here is the number of lines required: 36
