+++ tests/script_maps/valid_moves.txt
@@ -1,0 +1,1 @@
+DONE
