+++ tests/test_valid_output/Makefile
@@ -1,0 +1,14 @@
+# **************************************************************************** #
+#                                                                              #
+#                                                         :::      ::::::::    #
+#    Makefile                                           :+:      :+:    :+:    #
+#                                                     +:+ +:+         +:+      #
+#    By: ksuomala <ksuomala@student.hive.fi>        +#+  +:+       +#+         #
+#                                                 +#+#+#+#+#+   +#+            #
+#    Created: 2021/04/27 19:49:44 by ksuomala          #+#    #+#              #
+#    Updated: 2021/04/27 20:16:23 by ksuomala         ###   ########.fr        #
+#                                                                              #
+# **************************************************************************** #
+
+all:
+	gcc -Wall -Werror -Wextra -g *.c -I ../../libft -L ../../libft -lft -o tester
